Move the asv spin command into .spin/asv.py
The benchmark command and its helpers had grown to 142 lines, 27% of
.spin/cmds.py, alongside the unrelated docs, ipython, sdist and test
commands. They now have their own module, which spin loads through the
same [tool.spin.commands] entry.

ruff dropped a redundant local `import json` in _is_wheel_install while
reformatting; the module already imports it.
