+++ pyproject.toml
@@ -148,7 +148,7 @@ Environments = [
 ]
 Documentation = ['.spin/cmds.py:docs']
 Metrics = [
-    '.spin/cmds.py:asv'
+    '.spin/asv.py:asv'
 ]
 
 [tool.ruff.format]
